updated code
-space ship change
changed the temp image to the dedicated space ship

-movement
fixed movement and made it so that all keys can be enabled and disabled through the tempMovement(KEY) boolean.  Speed can also be changed.

-screen
changed the screen class to include and use the new changes and made the spaceship start at 0Xmiddle.